- move into cpp folder - float number on file/folder size - tree options to save some RAM

+++ cpp/include/utils.hpp
@@ -10,11 +10,11 @@
     extern const std::array<std::string_view, 6> sizes;
 
     struct size_unit {
-      size_t size;
+      float size;
       const std::string_view* unit;
     };
 
-    size_unit readable_file_size(size_t size_byte);
+    size_unit readable_file_size(float size_byte);
     size_unit readable_dir_size(fs::directory_entry path);
   }
 #endif
